pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "tradchem"
dynamic = ["version"]
description = "A comprehensive Python package for traditional medicine chemical database and analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "SaltyHeart", email = "contributors@tradchem.org"}
]
maintainers = [
    {name = "SaltyHeart", email = "contributors@tradchem.org"}
]
keywords = ["traditional medicine", "chemistry", "database", "SMILES", "cheminformatics", "ayurveda", "herbal medicine", "python package"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Natural Language :: English",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0,<2.0.0",
    "pandas>=1.3.0,<2.0.0",
    "scipy>=1.7.0,<2.0.0",
    "matplotlib>=3.4.0,<4.0.0",
    "seaborn>=0.11.0,<1.0.0",
    "scikit-learn>=0.24.0,<2.0.0",
    "jsonschema>=3.2.0",
    "pydantic>=1.8.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0,<8.0.0",
    "pytest-cov>=2.12.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0",
]
analysis = [
    "rdkit-pypi>=2022.9.1",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
]

[project.urls]
Homepage = "https://github.com/INSTITUTE-OF-SCIENTIFIC-INFORMATICS/Trad-Chem"
Documentation = "https://tradchem.readthedocs.io/"
Repository = "https://github.com/INSTITUTE-OF-SCIENTIFIC-INFORMATICS/Trad-Chem"
"Bug Tracker" = "https://github.com/INSTITUTE-OF-SCIENTIFIC-INFORMATICS/Trad-Chem/issues"

[project.scripts]
tradchem = "tradchem.cli:main"

[tool.setuptools]
packages = ["tradchem"]

[tool.setuptools.package-data]
tradchem = ["data/*.json"]

[tool.setuptools.dynamic]
version = {attr = "tradchem.version.__version__"}

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tradchem/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=tradchem",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["tradchem"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 